Task completed

+++ src/main/java/com/systemsmart/factory/RoomFactory.java
@@ -3,7 +3,7 @@ package com.systemsmart.factory;
 import com.systemsmart.entity.Room;
 
 public class RoomFactory {
-    public static Room createRoom(String type, int roomNumber, double price, boolean status) {
+    public static Room createRoom(String type, long roomNumber, double price, boolean status) {
         return new Room.Builder().setRoomNumber(roomNumber)
                 .setType(type).setPrice((int) price)
                 .setStatus(status).build();
